Xext: sync: move it into it's own subdir
