+++ Xext/sync/meson.build
@@ -1,0 +1,9 @@
+srcs_sync = [
+    'sync.c',
+]
+
+libxserver_sync = static_library('xserver_sync',
+    srcs_sync,
+    include_directories: inc,
+    dependencies: common_dep,
+)
