+++ test/xtest.c
@@ -41,7 +41,7 @@
 #include "exevents.h"
 #include "xkbsrv.h"
 #include "xserver-properties.h"
-#include "syncsrv.h"
+#include "Xext/sync/syncsrv.h"
 #include "tests-common.h"
 
 /**
